refactor: unify API for shim layers for the backends
